+++ docs/CHANGELOG.md
@@ -10,7 +10,7 @@ Version 1.5.0 (TBD): Not yet released
 - Restore cut first person weapon aim sway, toggleable with `cl_weaponsway`
 
 [@nickalreadyinuse](https://github.com/nickalreadyinuse)
-- Add `ui_con_color` console command to set the console background color
+- Add `ui_color_console` console command to set the console background color
 
 ### Bug fixes